libs/progressbar: Move into internal
