+++ internal/progressbar/progressbar_test.go
